requirements.txt
django>=5.1,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<26.0
gunicorn>=22.0,<26.0
